Make the source buildable with tools on Homebrew
Previously the souce can be built on the libraries that IBM provides and
special makefile, Makefile.osx, is required. And finally pkg is created.

But this method is removed when the LTFS project is published in GitHub.

This change makes this project buildable using tolls and libraried on Homebrew
with atandard GNU build method, configure, make, make install.

After this change, osxfuse, libxml2 and icu4c must be installed from homebrew.