Make lin_tape driver optional
To support wider platforms, I decided to switch default backend from
lin_tape to sg, because lin_tape only supports a few platform.

After this change, sg backend is the default in Linux and iokit backend
is the default in OSX. lin_tape backend is built only when
--enable-lintape is specified in configure script. And lin_tape backend
is set default when DEFAULT_TAPE=lin_tape is specified in configure
script.