Tiny fix of plugin help print (#36)
Fix to print plugin help correctly. It is broken when teh tape backend name prefix from "driver" to "tape".

This fix is a part of the freebsd patch provided by Spectra Logic. I will backport this portion to v2.4.0 branch. The freebsd patch will be merged soon in master branch.