Remove PACKAGE_XXX from ltfs.h (#120)
PACKAGE_NAME and PACKAGE_VERSION is defined for makefile build of macOS
but it is not a good idea to change multiple place when version number
is changed.

This change expects that value is provided from Malefile itself only for
macOS build from Makefile. There is no change for HomeBrew build.