Fix the bug of detecting read length
When READ6 is issued with sili the sg backend returns a wrong length.
This behavior doesn't create any error when LBP is disabled (default),
but LBP is enabled by `-o scsi_lbprotection=on` is specified in the
argument. The sg backend returns CRC error because wrong read length is
returned from _cdb_read().

The problem is the sg backend returns req.dxfer_len even if sili is
true. This field is [I] in the document so we must not expect the field
is updated in SG_IO ioctl. We must use req.resid instead.