Relax the index parser to allow 0-byte name (#182)
Name of root directory is treated as a volume name. So empty value is allowed when there is no volume name.

The current code only accepts <name/> as empty tag. So the index parser may reject an index written by another implementation.

This fix accepts both <name/> and <name></name>